requirements.txt
aiohttp
aiosqlite
feedparser
PyYAML
beautifulsoup4
python-dotenv
APScheduler